chatbot: remove dead abstractions (ChatLock interface, AwaitingInput outcome)
