+++ chatbot/src/main/java/pt/estga/chatbot/context/HandlerOutcome.java
@@ -3,7 +3,6 @@ package pt.estga.chatbot.context;
 public sealed interface HandlerOutcome
         permits HandlerOutcome.Success,
                 HandlerOutcome.Failure,
-                HandlerOutcome.AwaitingInput,
                 HandlerOutcome.Redispatch,
                 HandlerOutcome.StartNew,
                 HandlerOutcome.StartVerification {
@@ -11,8 +10,6 @@ public sealed interface HandlerOutcome
     record Success() implements HandlerOutcome {}
 
     record Failure() implements HandlerOutcome {}
-
-    record AwaitingInput() implements HandlerOutcome {}
 
     record Redispatch() implements HandlerOutcome {}
